Use patched vf2_sub_graph_iso.hpp.
The vf2_sub_graph_iso implementation from boost checks the
BidirectionalGraph concept. However this concept is not implemented by
the undirected adjacency matrix (don't ask why). The degree function is
not present. However, the vf2 implementation doesn't it. The concept
check is useless here. So we removed it in our version.